% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs.R
\name{checkSharedLibs}
\alias{checkSharedLibs}
\title{Check Shared Libraries}
\usage{
checkSharedLibs(vec, db)
}
\arguments{
\item{vec}{A (optional) character vector of shared libraries,
typcially with major soname. If missing a default vector is
used as fallback.}

\item{db}{An optional identifier for a library package database,
should be one of "18.04" or "19.10".}
}
\value{
Nothing currently but information is printed.
}
\description{
Check shared libraries on current system
}
\details{
This function examine the shared libraries in argument
\code{vec}. It runs \code{ldd} on these, filters for \dQuote{not
found} to identify unresolved shared libraries and tries to map
these against (stored) information on distribution packages (where
Ubuntu 18.04 and 19.10 are currently supported).

The internal helper function \code{.libs()} can be used to
identify shared libraries from packages in the current
\code{.libPaths()}.  Because search all packages in each 
\code{.libPaths()} entry at once (using shell tools) it does
not currently associate finds with the packages they originate
from.
}
\examples{
checkSharedLibs(c("libxml2.so.2"))
}
\author{
Dirk Eddelbuettel
}
